+++ src/services/Emails.php
@@ -177,6 +177,8 @@ class Emails extends Component
             'recipientType' => $email->recipientType,
             'to' => $email->to,
             'bcc' => $email->bcc,
+            'cc' => $email->cc,
+            'replyTo' => $email->replyTo,
             'enabled' => (bool)$email->enabled,
             'templatePath' => $email->templatePath,
             'attachPdf' => (bool)$email->attachPdf,
@@ -216,6 +218,8 @@ class Emails extends Component
             $emailRecord->recipientType = $data['recipientType'];
             $emailRecord->to = $data['to'];
             $emailRecord->bcc = $data['bcc'];
+            $emailRecord->cc = $data['cc'];
+            $emailRecord->replyTo = $data['replyTo'];
             $emailRecord->enabled = $data['enabled'];
             $emailRecord->templatePath = $data['templatePath'];
             $emailRecord->attachPdf = $data['attachPdf'];
@@ -358,7 +362,7 @@ class Emails extends Component
 
                 $newEmail->setTo($emails);
             } catch (\Exception $e) {
-                $error = Plugin::t( 'Email template parse error for custom email “{email}” in “To:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+                $error = Plugin::t('Email template parse error for custom email “{email}” in “To:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
                     'email' => $email->name,
                     'order' => $order->getShortNumber(),
                     'message' => $e->getMessage(),
@@ -375,7 +379,7 @@ class Emails extends Component
         }
 
         if (!$newEmail->getTo()) {
-            $error = Plugin::t( 'Email error. No email address found for order. Order: “{order}”', ['order' => $order->getShortNumber()]);
+            $error = Plugin::t('Email error. No email address found for order. Order: “{order}”', ['order' => $order->getShortNumber()]);
             Craft::error($error, __METHOD__);
 
             Craft::$app->language = $originalLanguage;
@@ -395,7 +399,55 @@ class Emails extends Component
                     $newEmail->setBcc($bcc);
                 }
             } catch (\Exception $e) {
-                $error = Plugin::t( 'Email template parse error for email “{email}” in “BCC:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+                $error = Plugin::t('Email template parse error for email “{email}” in “BCC:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+                    'email' => $email->name,
+                    'order' => $order->getShortNumber(),
+                    'message' => $e->getMessage(),
+                    'file' => $e->getFile(),
+                    'line' => $e->getLine()
+                ]);
+                Craft::error($error, __METHOD__);
+
+                Craft::$app->language = $originalLanguage;
+                $view->setTemplateMode($oldTemplateMode);
+
+                return false;
+            }
+        }
+
+        // CC:
+        if ($email->cc) {
+            try {
+                $cc = $view->renderString($email->cc, $renderVariables);
+                $cc = str_replace(';', ',', $cc);
+                $cc = preg_split('/[\s,]+/', $cc);
+
+                if (array_filter($cc)) {
+                    $newEmail->setCc($cc);
+                }
+            } catch (\Exception $e) {
+                $error = Plugin::t('Email template parse error for email “{email}” in “CC:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+                    'email' => $email->name,
+                    'order' => $order->getShortNumber(),
+                    'message' => $e->getMessage(),
+                    'file' => $e->getFile(),
+                    'line' => $e->getLine()
+                ]);
+                Craft::error($error, __METHOD__);
+
+                Craft::$app->language = $originalLanguage;
+                $view->setTemplateMode($oldTemplateMode);
+
+                return false;
+            }
+        }
+
+        if ($email->replyTo) {
+            // Reply To:
+            try {
+                $newEmail->setReplyTo($view->renderString($email->replyTo, $renderVariables));
+            } catch (\Exception $e) {
+                $error = Plugin::t('Email template parse error for email “{email}” in “ReplyTo:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
                     'email' => $email->name,
                     'order' => $order->getShortNumber(),
                     'message' => $e->getMessage(),
@@ -415,7 +467,7 @@ class Emails extends Component
         try {
             $newEmail->setSubject($view->renderString($email->subject, $renderVariables));
         } catch (\Exception $e) {
-            $error = Plugin::t( 'Email template parse error for email “{email}” in “Subject:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+            $error = Plugin::t('Email template parse error for email “{email}” in “Subject:”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
                 'email' => $email->name,
                 'order' => $order->getShortNumber(),
                 'message' => $e->getMessage(),
@@ -434,7 +486,7 @@ class Emails extends Component
         try {
             $templatePath = $view->renderString($email->templatePath, $renderVariables);
         } catch (\Exception $e) {
-            $error = Plugin::t( 'Email template path parse error for email “{email}” in “Template Path”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+            $error = Plugin::t('Email template path parse error for email “{email}” in “Template Path”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
                 'email' => $email->name,
                 'order' => $order->getShortNumber(),
                 'message' => $e->getMessage(),
@@ -451,7 +503,7 @@ class Emails extends Component
 
         // Email Body
         if (!$view->doesTemplateExist($templatePath)) {
-            $error = Plugin::t( 'Email template does not exist at “{templatePath}” which resulted in “{templateParsedPath}” for email “{email}”. Order: “{order}”.', [
+            $error = Plugin::t('Email template does not exist at “{templatePath}” which resulted in “{templateParsedPath}” for email “{email}”. Order: “{order}”.', [
                 'templatePath' => $email->templatePath,
                 'templateParsedPath' => $templatePath,
                 'email' => $email->name,
@@ -468,7 +520,7 @@ class Emails extends Component
         if ($email->attachPdf && $path = $email->pdfTemplatePath ?: Plugin::getInstance()->getSettings()->orderPdfPath) {
             // Email Body
             if (!$view->doesTemplateExist($path)) {
-                $error = Plugin::t( 'Email PDF template does not exist at “{templatePath}” for email “{email}”. Order: “{order}”.', [
+                $error = Plugin::t('Email PDF template does not exist at “{templatePath}” for email “{email}”. Order: “{order}”.', [
                     'templatePath' => $path,
                     'email' => $email->name,
                     'order' => $order->getShortNumber()
@@ -499,7 +551,7 @@ class Emails extends Component
                 $options = ['fileName' => $fileName . '.pdf', 'contentType' => 'application/pdf'];
                 $newEmail->attach($tempPath, $options);
             } catch (\Exception $e) {
-                $error = Plugin::t( 'Email PDF generation error for email “{email}”. Order: “{order}”. PDF Template error: “{message}” {file}:{line}', [
+                $error = Plugin::t('Email PDF generation error for email “{email}”. Order: “{order}”. PDF Template error: “{message}” {file}:{line}', [
                     'email' => $email->name,
                     'order' => $order->getShortNumber(),
                     'message' => $e->getMessage(),
@@ -519,7 +571,7 @@ class Emails extends Component
             $body = $view->renderTemplate($templatePath, $renderVariables);
             $newEmail->setHtmlBody($body);
         } catch (\Exception $e) {
-            $error = Plugin::t( 'Email template parse error for email “{email}”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
+            $error = Plugin::t('Email template parse error for email “{email}”. Order: “{order}”. Template error: “{message}” {file}:{line}', [
                 'email' => $email->name,
                 'order' => $order->getShortNumber(),
                 'message' => $e->getMessage(),
@@ -545,7 +597,7 @@ class Emails extends Component
             $this->trigger(self::EVENT_BEFORE_SEND_MAIL, $event);
 
             if (!$event->isValid) {
-                $error = Plugin::t( 'Email “{email}”, for order "{order}" was cancelled by plugin.', [
+                $error = Plugin::t('Email “{email}”, for order "{order}" was cancelled by plugin.', [
                     'email' => $email->name,
                     'order' => $order->getShortNumber()
                 ]);
@@ -559,7 +611,7 @@ class Emails extends Component
             }
 
             if (!Craft::$app->getMailer()->send($newEmail)) {
-                $error = Plugin::t( 'Commerce email “{email}” could not be sent for order “{order}”.', [
+                $error = Plugin::t('Commerce email “{email}” could not be sent for order “{order}”.', [
                     'email' => $email->name,
                     'order' => $order->getShortNumber()
                 ]);
@@ -572,7 +624,7 @@ class Emails extends Component
                 return false;
             }
         } catch (\Exception $e) {
-            $error = Plugin::t( 'Email “{email}” could not be sent for order “{order}”. Error: {error} {file}:{line}', [
+            $error = Plugin::t('Email “{email}” could not be sent for order “{order}”. Error: {error} {file}:{line}', [
                 'error' => $e->getMessage(),
                 'file' => $e->getFile(),
                 'line' => $e->getLine(),
@@ -650,6 +702,8 @@ class Emails extends Component
                 'emails.recipientType',
                 'emails.to',
                 'emails.bcc',
+                'emails.cc',
+                'emails.replyTo',
                 'emails.enabled',
                 'emails.templatePath',
                 'emails.attachPdf',